stupid gcc c99 compat

+++ src/util/frames.c
@@ -159,7 +159,7 @@ util_frames_t util_frames_await(util_frames_t frames)
   // need more to complete inbox
   if(frames->cache) return frames;
   // outbox is complete, awaiting flush
-  if((frames->out * PAYLOAD(frames)) > lob_len(frames->outbox)) return frames;
+  if((frames->out * PAYLOAD(frames)) > (uint32_t)lob_len(frames->outbox)) return frames;
   return NULL;
 }
